+++ src/features/posts/initialization/model/hooks.ts
@@ -1,0 +1,12 @@
+import { useEffect } from 'react';
+import { useTagsStore } from '../../../../entities/tags/model/store';
+
+export const usePostsInitialization = () => {
+  const { fetchTags: fetchTagsFromStore } = useTagsStore();
+
+  useEffect(() => {
+    fetchTagsFromStore();
+  }, [fetchTagsFromStore]);
+
+  return {};
+};
